+++ src/steps/cloud-asset/index.ts
@@ -147,6 +147,7 @@ export async function fetchIamBindings(
       if (permissions?.length ?? 0 > maxPermissionLength) {
         maxPermissionLength = permissions?.length ?? 0;
       }
+
       const conditionLength =
         (JSON.stringify(binding.condition) ?? '')?.length ?? 0;
       if (conditionLength > maxConditionLength) {
@@ -174,12 +175,7 @@ export async function fetchIamBindings(
   // Need to get bindings for every level of the Resource Hierarchy.
   try {
     // Project level bindings and all resource level bindings in that project
-    if (context.instance.config.projectId) {
-      await client.iterateIamPoliciesForProjectAndResources(
-        context.instance.config.projectId,
-        handlePolicyResult,
-      );
-    }
+    await client.iterateIamPoliciesForProjectAndResources(handlePolicyResult);
     // Folder level bindings
     await jobState.iterateEntities(
       { _type: FOLDER_ENTITY_TYPE },